% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse2}
\alias{collapse2}
\alias{collapse2.default}
\alias{collapse2.AbstractFArray}
\title{Collapse an array along certain margins}
\usage{
collapse2(x, along, method = c("sum", "average"), ...)

\method{collapse2}{default}(x, along, method = c("sum", "average"), ...)

\method{collapse2}{AbstractFArray}(x, along, method = c("sum", "average"), path = tempfile(), ...)
}
\arguments{
\item{x}{an array, a matrix, or a \code{\link{farray}} instance}

\item{along}{margin of the array}

\item{method}{whether to calculate summation or average, choices are
\code{'sum'} and \code{'average'}.}

\item{...}{passed to other methods.}

\item{path}{where to store the results; default is \code{\link[=tempfile]{tempfile()}}; see \link{farray}}
}
\value{
A collapsed array. For \code{\link{farray}}, returns a \code{farray} instance.

Collapsed array in \link{farray} format.
}
\description{
Collapse an array by calculating summation or average
along certain dimensions.
}
\details{
For matrices, we often calculate \code{\link{colSums}}, \code{\link{colMeans}} (collapse by
first margin), or \code{\link{rowSums}}, \code{\link{rowMeans}} (collapse by second margin). For
arrays, the collapse operation can be much more flexible than matrices.

Suppose array \code{x} has dimension \verb{100x200x300}, then collapse along the
first margin results in \verb{200x300} matrix, and collapse along the first two
margins results in a vector of length \code{300}.

For regular arrays, the dimensions of collapsed results equal to
\code{length(dim(x))-length(along)}. For \code{farray}s, the result dimension
will be at least two.
}
\examples{


# ------------------------------ Example: matrix ---------------------------
x <- array(1:10, c(2,5))

collapse2(x, 1)   # identical to colSums(x)

collapse2(x, 2)   # identical to rowSums(x)

collapse2(x, c(1,2)) # identical to sum(x)

# ----------------------------------- farray -------------------------------

x <- as.farray(array(1:100, c(2,5,10)))
x

collapse2(x, 1)

collapse2(x, c(1,3))


}
\author{
Zhengjia Wang
}
