% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{farray-threads}
\alias{farray-threads}
\alias{set_lazy_threads}
\alias{get_lazy_threads}
\title{Set Number of Threads for Lazy Arrays}
\usage{
set_lazy_threads(
  nr_of_threads = getOption("farray.nthreads"),
  reset_after_fork = NULL
)

get_lazy_threads(max = FALSE)
}
\arguments{
\item{nr_of_threads}{number of CPU cores to use, or \code{NULL} to
stay unchanged, default is \code{getOption('farray.nthreads')}}

\item{reset_after_fork}{whether to reset after forked process}

\item{max}{whether return maximum available threads}
}
\value{
Number of cores currently used.
}
\description{
Set number of threads used by 'OpenMP'
}
