% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorprod.R
\name{tensorprod}
\alias{tensorprod}
\title{Product of two arrays}
\usage{
tensorprod(
  A,
  B,
  alongA = integer(0),
  alongB = integer(0),
  path = tempfile(),
  meta_name = "farray.meta"
)
}
\arguments{
\item{A, B}{an \code{\link{array}} or \code{\link{farray}}}

\item{alongA, alongB}{integers; margins in \code{A} and \code{B} to be collapsed.}

\item{path, meta_name}{where to store the results; see \code{\link{farray}}.}
}
\value{
A \code{\link{farray}} object.
}
\description{
Consistent with \code{tensor()} function in \code{tensor} package, or
\code{ttl()} in \code{rTensor} package.
}
\details{
It is highly recommended that \code{A} is a small array and \code{B} is a
large array. The performance if highly related to how \code{alongB} is assigned.
If all the fast margins (see \code{\link{is_fast_margin}}) are contained within
\code{alongB}, then there is no memory optimization, and the whole calculation
occurs in the memory. If at least one fast margin is not in \code{alongB}, then
the calculation is memory optimized.
}
\examples{

x <- matrix(1:12, ncol = 4)
y <- array(rnorm(120), c(4,5,6))

# Equivalent to tensor::tensor(x, y, alongA = 2, alongB = 1)
# or rTensor::ttm(rTensor::as.tensor(y), x, m = 1)
tensorprod(x, y, alongA = 2, alongB = 1)

}
