% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{typeof,AbstractFArray-method}
\alias{typeof,AbstractFArray-method}
\title{Type of \code{farray}}
\usage{
\S4method{typeof}{AbstractFArray}(x)
}
\arguments{
\item{x}{a \code{farray} or an R object}
}
\value{
The storage type of data stored in the input
}
\description{
Type of \code{farray}
}
